backend/requirements.txt
Django==5.2.12                          # A high-level Python web framework that encourages rapid development and clean, pragmatic design.
dj-database-url==3.1.2                  # Use Database URLs in your Django Application.
djangorestframework==3.16.1             # Web APIs for Django, made easy.
djangorestframework-simplejwt==5.5.1    # A minimal JSON Web Token authentication plugin for Django REST Framework
dj-rest-auth==7.2.0                     # Authentication and Registration in Django Rest Framework
drf-spectacular==0.29.0                 # OpenAPI 3 schema generation for Django REST framework
psycopg==3.3.3                          # PostgreSQL database adapter for Python
Pillow==11.1.0                          # Python Imaging Library
